refactor: move endpoints services outside of /routes dir to /services/Endpoints
