+++ api/server/services/Endpoints/gptPlugins/initializeClient.spec.js
@@ -1,12 +1,12 @@
 // gptPlugins/initializeClient.spec.js
+const { PluginsClient } = require('~/app');
 const initializeClient = require('./initializeClient');
-const { PluginsClient } = require('../../../../app');
-const { getUserKey } = require('../../../services/UserService');
+const { getUserKey } = require('../../UserService');
 
 // Mock getUserKey since it's the only function we want to mock
-jest.mock('../../../services/UserService', () => ({
+jest.mock('~/server/services/UserService', () => ({
   getUserKey: jest.fn(),
-  checkUserKeyExpiry: jest.requireActual('../../../services/UserService').checkUserKeyExpiry,
+  checkUserKeyExpiry: jest.requireActual('~/server/services/UserService').checkUserKeyExpiry,
 }));
 
 describe('gptPlugins/initializeClient', () => {
