+++ api/server/services/Endpoints/openAI/buildOptions.js
